优化初始界面UI操作，制作添加血条、子弹数量UI
1. 将EventSystem放入Root中所有场景共用
2. 为所有UI添加每次进入都可以执行某些动作的函数FirstInit()
3. 制作血条和武器子弹数量UI，在svn添加相应资源
4. 在SpaceBattle场景添加新的血条，其他场景后续添加
5. 添加GamingUIManager控制游戏进行中的UI
6. 不小心改了下陨石的伤害，应该没啥问题
7. 删除TitleScene场景
8. 将项目设置中的QualitySetting改为只有Fantastic模式

+++ Assets/SpaceBattle/Scripts/Hinder.cs
@@ -9,7 +9,7 @@ public class Hinder : MonoBehaviour
     public GameObject explosion;
     //public GameObject stone;
     public float tumble = 1.0f;
-    public float damage = -40.0f;
+    public int damage = -40;
     public static Transform cam;
 
     public void Boom()
